Rename clib atomic to stdatomic and make it more compiant to the standard
