+++ auxlib/src/main/scala/scala/runtime/Statics.scala
@@ -1,7 +1,7 @@
 package scala.runtime
 
-import scala.scalanative.libc.atomic._
-import scala.scalanative.libc.atomic.memory_order._
+import scala.scalanative.libc.stdatomic._
+import scala.scalanative.libc.stdatomic.memory_order._
 
 /** Not for public consumption. Usage by the runtime only.
  */
