+++ clib/src/main/scala-2/scala/scalanative/libc/stdatomicExt.scala
@@ -1,0 +1,6 @@
+package scala.scalanative.libc
+
+import scala.scalanative.unsafe.extern
+
+@extern
+trait stdatomicExt { self: stdatomic.type => }
