+++ javalib/src/main/scala/java/util/concurrent/atomic/AtomicIntegerArray.scala
@@ -10,8 +10,8 @@ import scala.annotation.tailrec
 import scala.language.implicitConversions
 import scala.scalanative.annotation.alwaysinline
 import scala.scalanative.unsafe._
-import scala.scalanative.libc.atomic.CAtomicInt
-import scala.scalanative.libc.atomic.memory_order._
+import scala.scalanative.libc.stdatomic.AtomicInt
+import scala.scalanative.libc.stdatomic.memory_order._
 import java.util.function.IntBinaryOperator
 import java.util.function.IntUnaryOperator
 import scala.scalanative.runtime.IntArray
@@ -24,8 +24,8 @@ class AtomicIntegerArray extends Serializable {
   private[concurrent] def nativeArray: IntArray = array.asInstanceOf[IntArray]
 
   @alwaysinline
-  private implicit def ptrIntToAtomicInt(ptr: Ptr[Int]): CAtomicInt =
-    new CAtomicInt(ptr)
+  private implicit def ptrIntToAtomicInt(ptr: Ptr[Int]): AtomicInt =
+    new AtomicInt(ptr)
 
   /** Creates a new AtomicIntegerArray of the given length, with all elements
    *  initially zero.
