+++ scalalib/overrides-2.12/scala/collection/immutable/VM.scala
@@ -1,7 +1,7 @@
 package scala.collection.immutable
 
-import scala.scalanative.libc.atomic._
-import scala.scalanative.libc.atomic.memory_order._
+import scala.scalanative.libc.stdatomic._
+import scala.scalanative.libc.stdatomic.memory_order._
 
 // Backport from scala.runtime moved into s.c.immutable and made package
 // private to avoid the need for MiMa whitelisting.
