+++ scalalib/overrides-2.13.8/scala/collection/concurrent/TrieMap.scala.patch
@@ -54,14 +54,14 @@
 +  import scala.scalanative.runtime.RawPtr
 +  import scala.scalanative.unsafe.Ptr
 +  import scala.scalanative.annotation.alwaysinline
-+  import scala.scalanative.libc.atomic.memory_order.memory_order_release
-+  import scala.scalanative.libc.atomic.CAtomicRef
++  import scala.scalanative.libc.stdatomic.memory_order.memory_order_release
++  import scala.scalanative.libc.stdatomic.AtomicRef
  
 +  private class IntrinsicAtomicReferenceFieldUpdater[
 +      T <: AnyRef,
 +      V <: AnyRef
 +  ](@alwaysinline selector: T => Ptr[V]) extends AtomicReferenceFieldUpdater[T, V]() {
-+    @alwaysinline def atomicRef(obj: T) = new CAtomicRef(selector(obj))
++    @alwaysinline def atomicRef(obj: T) = new AtomicRef(selector(obj))
 +    @alwaysinline def compareAndSet(obj: T, expect: V, update: V): Boolean = atomicRef(obj).compareExchangeStrong(expect, update)
 +    @alwaysinline def weakCompareAndSet(obj: T, expect: V, update: V): Boolean =  atomicRef(obj).compareExchangeWeak(expect, update)
 +    @alwaysinline def set(obj: T, newIntalue: V): Unit = atomicRef(obj).store(newIntalue)
