+++ scripts/gyb_all.sh
@@ -36,8 +36,8 @@ gyb_files $nativelib $scala $unsafe Tag Nat CStruct CFuncPtr Size
 gyb_files $nativelib $scala $unsigned USize
 gyb_files $nativelib $scala $runtime Arrays Boxes Primitives
 
-gyb clib/src/main/scala/scala/scalanative/libc/atomic.scala.gyb
-gyb clib/src/main/resources/scala-native/atomic.c.gyb
+gyb clib/src/main/scala/scala/scalanative/libc/stdatomic.scala.gyb
+gyb clib/src/main/resources/scala-native/stdatomic.c.gyb
 
 gyb $javaNIO/HeapBuffers.scala.gyb
 gyb $javaNIO/HeapByteBufferViews.scala.gyb
@@ -45,4 +45,3 @@ gyb $javaNIO/MappedByteBufferViews.scala.gyb
 gyb $javaNIO/PointerByteBufferViews.scala.gyb
 
 gyb unit-tests/native/src/test/scala/org/scalanative/testsuite/niobuffer/ByteBufferViewsNativeTests.scala.gyb
-
